Rework golden testing (#6)
These tests reinvented a lot of existing technology for golden testing.
Switched to pytest and cleaned them up using that. In doing so, fixed
several tests that were silently failing and added
`torch.use_deterministic_algorithms` to the `make_deterministic` option,
which was required to get deterministic results on MI210.

Tests now have defined names, can be selected as pytest parametrized
tests, can fail fast, can display output or not, etc.

I didn't use either of the existing pytest golden plugins because
neither of them actually do what we want here. One requires tests to be
specified as yaml files and the other takes over file comparison.

There's still some weird parsing of bash files here, but I decided to
not mess with that too much as it appears to work at the moment.

It also still floods you with temporary directories in a way that seems
a bit unnecessary and could probably be managed by pytest, but the code
has a number of specific dependencies on exactly where a command is run
that I didn't want to touch right now.