Thread device through model initialization
The device kwarg here is a common, albeit relatively new pattern. Adding
it was a boring traversal of the call graph.

This allows creating models directly on device or even uninitialized.
This dramatically speeds up model creation, which matter most for
running the end-to-end tests, where the model is separately initialized
in a bunch of different processes, and even more especially when trying
to run those tests in parallel where torch will overwhelm the CPU.

This allows me to get almost full parallelism speedups on the example
script tests on multiple GPUs, whereas previously I was seeing dramatic
dropoff (6 GPUs was more than half the time of 1 GPU on a machine with
192 CPU cores).

Of course, with https://github.com/Brium/RFdiffusion/pull/17, that maybe
becomes unimportant, but I'd already done this and I think it's just
generally an improvement.

Theoretically this should allow us to skip initialization entirely using
`torch.nn.utils.skip_init` and just load the weights from a checkpoint,
but I found that in practice this did not speed things up (if anything,
it was slower). This is perhaps because RFdiffusion is doing some custom
initialization that doesn't play nicely with the "meta" device used by
`skip_init`. Given no large speedups, I don't think it's worth playing
with this more advanced feature.