Use BioPython for processing pdbs in tests
I'm not confident enough about the workings of everything else to
replace the `calc_rmsd` utils function everywhere, but at least for the
tests, processing with BioPython seems significantly more robust. It was
able to find more optimal transformations when superimposing proteins
for RMSD calculation. Even then though, it still finds a ~1e-15 RMSD
when two pdbs are byte-identical. It's nice to actually report 0 in this
case, so we now first calculate the naive RMSD without transformation
and use that if it's smaller. I suspect this will only come up in the
identical-file case, but it could also give a better RMSD for two
different pdbs that are already very close. With those changes, we're
able to see that there's actually no diff in output run to run.

Amongst other issues, the utils function adds an epsilon value to the
RMSD before taking the square root, which means it always reports
non-zero. I'm not sure why, since `np.sqrt` is perfectly happy with
inputs of zero. Even removing that though, BioPython still performs
better.

The disadvantage, of course, is that this adds a dependency.