Diffusion regression tests based on hydra config (#13)
The existing tests that run the examples have the advantage that they
can reuse something that already exists and they ensure the user-facing
examples don't break. But for just expanding test coverage, it's a lot
nicer and less error prone to write specs directly IMO. I want to use a
similar setup for benchmarking as well.