Remove direct usage of `opt_einsum` (#14)
Torch already uses this under the hood when you call `torch.einsum`, but
better and in a way that's compatible with `torch.compile`.

I swapped the argument order on a couple of einsums in order to preserve
bit-exact output. I was getting a discrepancy with the change of these
einsums to `torch.compile`. I traced through the code and if
`opt_einsum` is installed, `torch.einsum` calls it iff the einsum has
more than two inputs (otherwise there's no optimization to be done
AFAIU). On the other hand, `opt_einsum` if called directly in this case
does a bunch of fiddling around to not optimize the path and then calls
`torch.einsum`. Somewhere in that fiddling for some of the einsums,
`opt_einsum` seems to swap the order of the operands. So we get
differences in the output on the order of `5e-8` and that stacks up into
an overall difference in the RMSD for the protein. This also explains
why removing `opt_einsum` is in general a performance gain: torch
already reaps all the possible performance benefits but does it better.
So we can get identical output (at the PDB level) by just swapping the
order and calling torch directly.

This also raises the point though that there's nothing really magical
about the current numeric output. Either einsum order here is reasonable
(and if anything the version that calls `torch.einsum` directly without
swapping the args is arguably more correct, as its closer to the
programmer intent). It's really nice to have the ability to test and
confirm that there's no diff from any given change, but for some more
complicated things like `torch.compile` that's going to be quite
difficult to guarantee (and severely limits optimization opportunities).
The strict RMSD limit is probably something we will want to relax for
certain changes, therefore.

Max RMSD diff: 0.0