Fix worker idx without `pytest-xdist` (#16)
I foolishly didn't test this. I guess I forgot or thought *of course* a
try/except for any exception would work here. But defining the fixture
with the `worker_id` argument doesn't create the error, so that works
fine. It's when we try to use it that we hit the error.