Avoid device/host sync in `make_full_graph` (#18)
This avoids the device-host sync present in the original implementation,
which was also basically constructing an identity matrix in a strange
way. The issue is that `torch.nonzero` cannot now from shape information
alone what size the output is, so it requires a device/host sync. There
is also an (undocumented
https://github.com/pytorch/pytorch/issues/152347) torch.nonzero_static
method (https://github.com/pytorch/pytorch/issues/62320), but it breaks
`torch.compile` (https://github.com/pytorch/pytorch/issues/152634)
(configuration to use that coming soon).

So instead we make use of our knowledge of what is actually being
computed here. It relies on my analysis that the entries in `idx` are
all unique (otherwise there would be off-diagonal entries that are also
zero).

Max RMSD diff: 0.0