Script for benchmarking inference steps
This focuses on the step level because it allows us to get results much
faster and evaluate inference time separately from initialization
overhead. The idea is that this is likely actually a better proxy for
performance when creating many designs than running inference to design
completion. It's also about 50 times faster, which is nice for iterative
development.

I cleaned up a couple things in the regular inference run as well as I
was adapting it.