Correctly calculate sampler steps
Technically you could use a different end step although we deliberately
don't do that for benchmarking because it actually changes the output of
sample step and we want to simulate the middle of a run. Also because
calculating the correct end step to put in order to achieve the desired
number of warmup and benchmarking steps seems convoluted.