Translate the rest of the example scripts
Fixed lengths help a lot with determinism. I gave all the configs nice
names also. I took a look at the contig parsing code, and it turns out
that you can just list a single number when the contig is a fixed length
rather than a range, so "150" instead of "150-150":

https://github.com/Brium/RFdiffusion/blob/7cd3ac12b0/rfdiffusion/contigs.py#L44-L48