Fix logging setting
I was trying to figure out how to make pytest output logs interleaved
with stdout, as that would match the normal output when running
intference since RFdiffusion freely mixes logging and prints. I couldn't
get that to work though, so the best we can do is actually show the info
logs by default. If more of the RFdiffusion output were put into logs
instead of prints then it would be more shared again.