Still use contigmap even without contigs
I found this helpful for passing through the length and having that get
validated. I can drop it if it seems like a bad idea. I do think it's a
bit weird that we throw away the whole contigmap if contigs isn't set
here. Mabye the idea is that the rest of the parameters don't make sense
without contigs?