Clean up logging
RFdiffusion uses the Python logger, but also uses print statements a
bunch. This results in a mix of different output that is hard to deal
with in scripts and tests and such. This converts a bunch of print
statements to log messages instead. It doesn't quite get all of them,
but I think it's still an improvement.