Don't do unnecessary chain break finding (#36)
ae036e30c6 added code for macrocyclic peptides, which includes finding
sequence breaks. But this break-finding is performed even when it isn't
used. It is implemented with numpy and forces a device-host transfer.

This moves it below the conditional that actually uses the break
finding. This results in noticeable performance gains.